requirements.txt
asgiref==3.4.1
Django==3.2.8
django-cors-headers==3.10.0
django-csp==3.7
djangorestframework==3.12.4
gunicorn==20.1.0
python-dotenv==0.19.1
pytz==2021.3
sqlparse==0.4.2
typing-extensions==3.10.0.2
